Code review

+++ src/xrGame/ActorHelmet.cpp
@@ -10,7 +10,7 @@ CHelmet::CHelmet()
 {
     m_flags.set(FUsingCondition, TRUE);
     m_HitTypeProtection.resize(ALife::eHitTypeMax);
-    for (int i = 0; i < static_cast<int>(ALife::eHitTypeMax); i++)
+    for (u32 i = 0; i < ALife::eHitTypeMax; i++)
         m_HitTypeProtection[i] = 1.0f;
 
     m_boneProtection = xr_new<SBoneProtections>();
